Adds the bind-to-interface helper code
Typical UDP socket would listen on 0.0.0.0:67 and bind to an interface
to limit exposure. This PR implements a helper method to bind and serve
on a socket.

Currently it’s linux only. I left an unimplemented!() trap for BSD with
a note how to progress there and feature-gated the code.