Improve use of Rust types (#7)
* Rename Option to DhcpOption.

It's better not to cause confusion with the Rust std::options::Option type. We can syntactically disambiguate between these types, but it's still a bit confusing for users of the library to see `Option` used in this way.

* Simplify DHCP request handler function type.

Rather than extracting the message type from a Packet and then passing both into a request handler (conditional on the type being valid), just pass the Packet through unconditionally. This both simplifies request handling and allows handlers to make decisions about how to deal with erroneous requests (rather than silently dropping them on the floor).

* Convert DHCP message types into a Rust enum.

Also apply `rustfmt` changes to the lines being touched anyway.

* Add more documentation to MessageType enum.